src/app/shared/support/support.component.spec.ts
import { async, ComponentFixture, TestBed } from '@angular/core/testing';
import { BrowserModule, By } from '@angular/platform-browser';
import { FormsModule, ReactiveFormsModule } from '@angular/forms';
import { DebugElement } from '@angular/core';
import { HttpModule } from '@angular/http';

import { HttpClientModule } from '@angular/common/http';
import { RouterTestingModule } from '@angular/router/testing';
import { SupportComponent } from './support.component';

fdescribe('SupportComponent', () => {
  let component: SupportComponent;
  let fixture: ComponentFixture<SupportComponent>;
  let debug: DebugElement;
  let el: HTMLElement;

  beforeEach(async(() => {
    TestBed.configureTestingModule({
      declarations: [ SupportComponent ],
      imports: [
        BrowserModule,
        HttpClientModule,
        HttpModule,
        RouterTestingModule,
        FormsModule,
        ReactiveFormsModule
      ]
    })
    .compileComponents();
  }));

  beforeEach(() => {
    fixture = TestBed.createComponent(SupportComponent);
    component = fixture.componentInstance;
    debug=fixture.debugElement.query(By.css('form'));
    el=debug.nativeElement;
    fixture.detectChanges();
  });

  it('should create', () => {
    expect(component).toBeTruthy();
  });

  it('should render title in a h1 tag', async(() => {
    fixture.detectChanges();
    const e1=fixture.nativeElement.querySelector('h1');
    expect(e1.innerText.toLowerCase()).toEqual('support');
  }));

  it('should call the submitIssue method', async(() => {
    fixture.detectChanges();
    spyOn(component,'submitIssue');
    el=fixture.debugElement.query(By.css('button')).nativeElement;
    el.click();
    expect(component.submitIssue).toHaveBeenCalledTimes(0);
  }));

  it('form should be invalid', async(() => {
    component.rForm.controls['issueTitle'].setValue('');
    component.rForm.controls['issueDescription'].setValue('');
    component.rForm.controls['inputEmail'].setValue('');
    expect(component.rForm.valid).toBeFalsy();
  }));

  it('form should be valid', async(() => {
    component.rForm.controls['issueTitle'].setValue('Login');
    component.rForm.controls['issueDescription'].setValue('Not able to login');
    component.rForm.controls['inputEmail'].setValue('anu@123');
    expect(component.rForm.valid).toBeTruthy();
  }));
});
